backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
websockets>=13.0
apscheduler>=3.10.4
pydantic>=2.9.0
cachetools>=5.5.0
